refactor: usernamesetting -> userconfirm
